app file and readme updated
redistricting-app.py replaced with app.py with adjustements from deployment on PythonAnywhere (http://dadeda.pythonanywhere.com/).

Note PythonAnywhere is git initialized as a remote for this repo, but no longer contains all of the files; the Notebook folder was removed and other files not required for the deployment in order to limit the disk space there to stay under the free level for the time being.